requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7; python_version >= '3'
cryptography==35.0.0
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.45.0
django-cors-headers==3.10.0
django-rest-auth==0.9.5
django==3.2.8
djangorestframework==3.12.4
idna==3.3; python_version >= '3'
oauthlib==3.1.1; python_version >= '3.6'
pillow==8.4.0
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.3.0; python_version >= '3.6'
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
